requirements.txt
django
dj-database-url
djangorestframework
pillow
psycopg2-binary
django-cors-headers
djangorestframework_simplejwt
django-rest-swagger
WeasyPrint
drf-api-logger